Version 1 of the project is officially done.
I ran into an issue where the player sprite (the spaceship) was too wide, meaning that it reached the boundary of the screen before the middle of the ship, which is the origin point of the bullets, could align with the enemies closest to the edge, making those enemies unkillable. I just got rid of the ship sprite altogether and replaced it with a Dr. Mario head, since I had planned to do that anyway to match the enemies and bullets thematically. This sprite is 24x24 pixels just like the enemies, making things nice and uniform, if a bit small.

When an enemy reaches the bottom of the screen, it is removed from the game and player lives are reduced by 1. The name of the game, the current score and the remaining lives are all displayed as the window title. Note that this is not really much of a game at this point, since reaching 0 lives does absolutely nothing, and there is not actually any way to increase score, so it will always be at 0.

I could spend a little bit more time and add score increase when bullets kill enemies, but seeing as that was not one of the criteria for version 1, I will instead spend that time working on version 2.

Enemies have been sped up for testing purposes, and lives have been increased from the arbitrary number 10 to 72, which is currently the hardcoded number of enemies, so that I at least don't have to look at a negative lives count until I make an actual failure state for the game.